mechanical move of NewTemplate* methods from expander to util package
